Much simpler, and the output file isn't massive
