--- scripts/gulp-prop-data.js
@@ -1,69 +1,0 @@
-const docgen = require('react-docgen-typescript');
-const fs = require('fs');
-const through = require('through2');
-
-const config = {
-  savePropValueAsString: true,
-  propFilter: (prop, component) => {
-    if (prop.declarations !== undefined && prop.declarations.length > 0) {
-      const hasPropAdditionalDescription = prop.declarations.find((declaration) => {
-        return !declaration.fileName.includes('node_modules');
-      });
-
-      return Boolean(hasPropAdditionalDescription);
-    }
-
-    return true;
-  },
-};
-
-const customParser = docgen.withCustomConfig('./tsconfig.json', config);
-
-module.exports = function (propsFileName) {
-  if (typeof propsFileName !== 'string') {
-    throw new Error('Must provide a filename to gulp-prop-data');
-  }
-
-  const tsFilePaths = [];
-
-  function processTsFile(tsFile, enc, cb) {
-    tsFilePaths.push(tsFile.path);
-
-    cb();
-  }
-
-  function endStream(cb) {
-    const components = customParser.parse(tsFilePaths);
-    const componentMap = new Map();
-
-    // const writeStream = fs.createWriteStream(propsFileName, { flags: 'w' });
-    // writeStream.write('[\n');
-
-    for (const component of components) {
-      const data = {
-        displayName: component.displayName,
-        filePath: component.filePath,
-        props: Object.values(component.props).map(
-          ({ defaultValue, description, name, required, type }) => ({
-            defaultValue,
-            description,
-            name,
-            required,
-            type: type.name,
-          })
-        ),
-      };
-      // There are duplicate entries for each component because it looks at
-      // each export, but they should all be the same, so it doesn't matter
-      // if we write over previously set data.
-      componentMap.set(data.displayName, data);
-      // writeStream.write(JSON.stringify(data, null, 2) + ',\n');
-    }
-
-    // writeStream.write(']\n');
-    // writeStream.end();
-    cb();
-  }
-
-  return through.obj(processTsFile, endStream);
-};
